requirements.txt
fastapi==0.85.0
gunicorn
uvicorn[standard]==0.18.3
pandas==1.5.0
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
urllib3==1.26.12
charset-normalizer==2.1.1
pycryptodome==3.20.0
pyotp==2.7.0
ipinfo==4.3.1
pystache==0.6.0
pyqrcode==1.2.1
pypng==0.20220715.0
psycopg2-binary==2.9.4
pdfkit==1.0.0
wkhtmltopdf==0.2
aiosmtplib==2.0.2
python-magic==0.4.27
eth-account==0.11.0
web3==6.15.1